refactor: adjust directory structure and packet implementation of handler
